Implement preliminary wall collision (still buggy)

+++ src/game/Box.java
@@ -2,11 +2,11 @@ package game;
 
 import org.joml.Matrix4f;
 import org.joml.Vector2f;
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 
 public class Box extends GameObject {
-    public final Vector3f dimensions;
-    public Box(Vector3f position, Vector3f dimensions) {
+    public final Vector3d dimensions;
+    public Box(Vector3d position, Vector3d dimensions) {
         super(position);
         this.dimensions = dimensions;
     }
